requirements.txt
adafruit-ampy==1.1.0
alembic==1.13.3
bitarray==2.9.3
bitstring==4.2.3
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
esptool==4.8.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mysql-connector-python==9.0.0
mysqlclient==2.2.5
pycparser==2.22
PyMySQL==1.1.1
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
reedsolo==1.7.0
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.2.1
